// src/models/enums/jdSkillType.ts
export enum JdSkillType {
    Typescript = 'typescript',
    React = 'react',
    ReactNative = 'reactnative',
    MobX = 'mobx',
    Sass = 'sass',
    Css = 'css',
    Figma = 'figma',
    ProductDesign = 'productdesign',
    ReactNativeReanimated = 'reanimated',
    Redux = 'redux',
    NodeJs = 'nodejs',
    CSharp = 'c#',
    NoSql = 'nosql',
    Sql = 'sql',
    AfterEffects = 'aftereffects',
    MotionDesign = 'motiondesign',
    BrandDesign = 'branddesign',
}